ssa: include recover blocks in post dominance.
The previosuly disabled recover blocks for post dominace is enabled.
The fix is trivial since (assumption) there is only one recover block with no successors and predecessors.
As a result, it gets naturally accounted as an unreachable block and hence reachable from the synthetic exit block.
No special action is needed.